pricing/requirements.txt
flask
numpy
apscheduler
flask_socketio
gevent
gevent-websocket
python-dotenv
datetime